+++ src/model/ModuleMapping.h
@@ -61,7 +61,7 @@ public:
 
     double getMinimum() const;
     double getMaximum() const;
-    float getDeadbandValue() const;
+    double getDeadbandValue() const;
 
     int getAddress() const;
     LabelMap getLabelMap() const;
@@ -81,7 +81,7 @@ private:
 
     double m_minimum = 0.0;
     double m_maximum = 1.0;
-    float m_deadbandValue = 0.0;
+    double m_deadbandValue = 0.0;
 
     int m_address = -1;
     LabelMap m_labelMap{};
